backend_v2/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
pgvector
langchain
llama-index
pydantic
PyPDF2
python-dotenv
